Move Windows-specific path validation out of shared FileStream
FileStream currently has a few path validation checks that are specific to Windows. This commit moves these checks into a Windows-specific file.
